django/requirements.txt
aniso8601==9.0.1
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.6
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
drf-yasg==1.21.5
filelock==3.8.2
graphene==3.2.2
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
numpy==1.24.1
packaging==23.0
platformdirs==2.6.0
promise==2.3
protobuf==4.21.12
pytz==2022.7.1
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.17.1
